refactor: update import paths for sensor models
