+++ api/src/main.py
@@ -11,7 +11,7 @@ from sqlmodel import Session, select
 
 from .config import settings
 from .database import create_db_and_tables, get_session
-from .models import Sensor, SensorHistory
+from .models.sensor_model import Sensor, SensorHistory
 from .services.scheduler_service import scheduler
 
 # Initialize Sentry SDK
