+++ api/src/models/sensor_model.py
