+++ api/tests/test_sensor_service.py
@@ -8,7 +8,7 @@ from src.database import (
     create_engine,
     drop_db_and_tables,
 )
-from src.models import SensorHistory
+from api.src.models.sensor_model import SensorHistory
 from src.services.sensor_service import check_sensor_task
